requirement.txt
colorama
ollama
duckduckgo-search
ddgs
requests
chromadb
sentence-transformers
openai
google-genai
anthropic
dotenv

